Add RT-RRTstar ability to find goal state quickly
ompl: Add functionality to check if shortest path to goal is clear and
add it to the tree if it is. Need check_shortest_path parameter to be
set to true to do this. By default it is false.

panda_moveit_config: Add RT-RRTstar parameter to activate checking of
shortest path in `ompl_planning.yaml`

THES-27: Implement ability to find initial path to goal state quickly