Update to dynamic/monitored RRT* demo
end_effector_control: changes to the demo interface allow the demo to
react to goal state changes live, instead of having to wait for the
current path execution to complete as before. Also includes changes to
ensure the demo can be run in both simulation or on hardware.

THES-76, THES-25, THES-29